GPLIM-2768: Changed quality metric to switch off of Bass's data type instead of the aggregation's data type. Added null check to avoid NPE when displaying contamination. Changed RNA quality metric to use pfAlignedBases instead of pfReadsAligned. Made the metrics persistence context EXTENDED to avoid LazyInitializationExceptions when lazy-loading from DTO. Changed metrics query to be more specific to each Bass file. Added "Loading..." text to Submissions Requests tab while data is being loaded. Prevented reloading Submissions Requests tab when switching to another tab and back (only after tab content has been successfully loaded; otherwise the request is canceled).

+++ mercury/src/main/java/org/broadinstitute/gpinformatics/infrastructure/metrics/entity/AggregationAlignment.java
@@ -27,8 +27,8 @@ public class AggregationAlignment implements Serializable {
     private int aggregationId;
     @Id
     private String category;
-    @Column(name = "PF_READS_ALIGNED") private Integer pfReadsAligned;
-    @Column(name = "PF_ALIGNED_BASES") private Integer pfAlignedBases;
+    @Column(name = "PF_READS_ALIGNED") private Long pfReadsAligned;
+    @Column(name = "PF_ALIGNED_BASES") private Long pfAlignedBases;
     @ManyToOne
     @JoinColumn(name = "AGGREGATION_ID", referencedColumnName = "ID", nullable = false)
     private Aggregation aggregation;
@@ -38,7 +38,7 @@ public class AggregationAlignment implements Serializable {
      */
     protected AggregationAlignment() {}
 
-    public AggregationAlignment(Integer pfReadsAligned) {
+    public AggregationAlignment(Long pfReadsAligned) {
         this.pfReadsAligned = pfReadsAligned;
     }
 
@@ -46,7 +46,7 @@ public class AggregationAlignment implements Serializable {
         return category;
     }
 
-    public Integer getPfReadsAligned() {
+    public Long getPfReadsAligned() {
         return pfReadsAligned;
     }
 
@@ -84,7 +84,7 @@ public class AggregationAlignment implements Serializable {
         return true;
     }
 
-    public Integer getPfAlignedBases() {
+    public Long getPfAlignedBases() {
         return pfAlignedBases;
     }
